requirements.txt
pyheif
Pillow
